requirements.txt
kivy==2.2.1
requests
pillow
kivymd
google-generativeai
python-dotenv
networkx
matplotlib
reportlab
python-docx
putergenai
aiohttp
thefuzz
fuzzywuzzy
python-Levenshtein
pydub
simpleaudio
pyinputplus
pyperclip
pyinstaller
